outsource parser prototypes to parser.h
